requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pywebpush==1.14.0
apscheduler==3.10.4
python-dotenv==1.0.1
